# fix cardinality aggregation performance