--- a/src/aggregation/metric/cardinality.rs
+++ b/src/aggregation/metric/cardinality.rs
@@ -174,7 +174,7 @@ impl SegmentCardinalityCollector {
                 }
             }
             term_ids.sort_unstable();
-            dict.ords_to_term_cb(term_ids.iter().map(|term| *term as u64), |term| {
+            dict.sorted_ords_to_term_cb(term_ids.iter().map(|term| *term as u64), |term| {
                 self.cardinality.sketch.insert_any(&term);
             })?;
             if has_missing {