--- a/sstable/src/value/range.rs
+++ b/sstable/src/value/range.rs
@@ -78,6 +78,7 @@ impl ValueWriter for RangeValueWriter {
 }
 
 #[cfg(test)]
+#[allow(clippy::single_range_in_vec_init)]
 mod tests {
     use super::*;