--- a/sstable/src/vint.rs
+++ b/sstable/src/vint.rs
@@ -39,8 +39,6 @@ pub fn deserialize_read(buf: &[u8]) -> (usize, u64) {
 
 #[cfg(test)]
 mod tests {
-    use std::u64;
-
     use super::{deserialize_read, serialize};
 
     fn aux_test_int(val: u64, expect_len: usize) {